requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
accelerate == 0.12.0
absl-py
ml_collections
einops
ftfy == 6.1.1
transformers == 4.23.1
-e clip @ git+https://github.com/openai/CLIP.git@main#egg=clip
xformers